backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19
alembic>=1.13
passlib[bcrypt]>=1.7
bcrypt<4.1
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
pillow>=10.2
pydantic-settings>=2.1
email-validator>=2.1
httpx>=0.27
pytest>=8.0
pytest-asyncio>=0.23
